+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Changelog
 
-## [Unreleased](https://github.com/schemathesis/schemathesis/compare/v4.18.3...HEAD) - TBD
+## [Unreleased](https://github.com/schemathesis/schemathesis/compare/v4.18.4...HEAD) - TBD
+
+## [4.18.4](https://github.com/schemathesis/schemathesis/compare/v4.18.3...v4.18.4) - 2026-05-12
 
 ### :bug: Fixed